sonar critical for errorhandling
Cast one of the operands of this multiplication operation to a long
Sonar Link:
https://sonar.onap.org/project/issues?id=org.onap.dmaap.messagerouter.msgrtr%3Amsgrtr&open=AV4-VakU32hFUzlqc5wn&resolved=false&types=BUG
Location:
src/main/java/com/att/nsa/cambria/service/impl/EventsServiceImpl.java
Line No-439

Issue-ID: DMAAP-353

+++ src/main/java/com/att/nsa/cambria/service/impl/EventsServiceImpl.java
@@ -436,7 +436,7 @@ public class EventsServiceImpl implements EventsService {
 		final long startMs = System.currentTimeMillis();
 		long count = 0;
 		
-		long maxEventBatch=1024 * 16;
+		long maxEventBatch=(long)1024 * 16;
 		String batchlen = AJSCPropertiesMap.getProperty(CambriaConstants.msgRtr_prop,BATCH_LENGTH);
 		if(null!=batchlen)maxEventBatch=Long.parseLong(batchlen);
